Boosted performance
Added a method which boosts performance by saving memory and time. In this method we keep track of only the last digit of the Fibonacci number. Usage of arrays is avoided. The last digit of a Fibonacci series repeats itself after the 60ith number and hence we check till n%60 further saving time.